Fix potential overflow

+++ java/src/kanzi/Global.java
@@ -182,7 +182,7 @@ public class Global
          return log << 10;
 
       long base = 0;
-      long z = x - (1 << log);
+      long z = (long) x - (1L << log);
 
       // Use the fact that log2(x) = log2(2^(log2(x)+1)*y) = log2(2^p)+ 1 + ln(1-z)/ln(2)
       // with z in ]0, 0.5[, it yields log2(x) = p + 1 - (z/1 + z^2/2 + z^3/3 ...)/ln(2)
@@ -195,7 +195,7 @@ public class Global
          if (z < 1 << (log-1) + 1 << (log-2)) 
          {
             base = 497;
-            x = (int) ((long) x * 5 / 7);   
+            x = (int) ((x * 5L) / 7L);   
          }
          
          // z in [0.75, 1[ => select 1 - x Taylor series expansion
